cpcgen: clang errors
/code/illumos-gate/usr/src/lib/libcustr/common/custr.c:195:17: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
        while (chunksz < len) {
               ~~~~~~~ ^ ~~~
1 error generated.

cpcgen.c:123:9: error: missing field 'cwhite_human' initializer [-Werror,-Wmissing-field-initializers]
        { NULL }
               ^
cpcgen.c:161:44: error: missing field 'cpapi_papi' initializer [-Werror,-Wmissing-field-initializers]
        { "MEM_INST_RETIRED.STORES" "PAPI_sr_ins" },
                                                  ^
cpcgen.c:162:43: error: missing field 'cpapi_papi' initializer [-Werror,-Wmissing-field-initializers]
        { "MEM_INST_RETIRED.LOADS" "PAPI_ld_ins" },
                                                 ^
cpcgen.c:1217:42: error: format specifies type 'int' but the argument has type 'pid_t' (aka 'long') [-Werror,-Wformat]
        if (asprintf(&tmpname, ".%s.%d", fname, getpid()) == -1) {
                                    ~~          ^~~~~~~~
                                    %ld
cpcgen.c:1286:42: error: format specifies type 'int' but the argument has type 'pid_t' (aka 'long') [-Werror,-Wformat]
        if (asprintf(&tmpname, ".%s.%d", fname, getpid()) == -1) {
                                    ~~          ^~~~~~~~
                                    %ld
cpcgen.c:1625:42: error: format specifies type 'int' but the argument has type 'pid_t' (aka 'long') [-Werror,-Wformat]
        if (asprintf(&tmpname, ".%s.%d", fname, getpid()) == -1) {
                                    ~~          ^~~~~~~~
                                    %ld
cpcgen.c:1834:42: error: format specifies type 'int' but the argument has type 'pid_t' (aka 'long') [-Werror,-Wformat]
                if (asprintf(&tmpname, ".%s.%d", name, getpid()) == -1) {
                                            ~~         ^~~~~~~~
                                            %ld
7 errors generated.

+++ usr/src/lib/libcustr/common/custr.c
@@ -192,7 +192,7 @@ custr_append_vprintf(custr_t *cus, const char *fmt, va_list ap)
 		return (len);
 	}
 
-	while (chunksz < len) {
+	while (chunksz < (size_t)len) {
 		chunksz *= 2;
 	}
